// src/enums/chain-contracts-addresses.enum.ts
export enum chainTicTacAvaxAddress {
	ARBITRUM_SEPOLIA = '0xc398B0Dc9035a852b2Cf614d6526d526A500b9C3',
	AVALANCHE_FUJI = '',
	BASE_SEPOLIA = 'base-sepolia',
	CELO_ALFAJORES = 'celo'
}

export enum chainTicTacAvaxCrossAddress {
	ARBITRUM_SEPOLIA = '0xeA6831B94eC51DFc8865f9a3AC8adD8ddf92C28F',
	AVALANCHE_FUJI = '',
	BASE_SEPOLIA = 'base-sepolia',
	CELO_ALFAJORES = 'celo'
}